# apps/backend/static/openapi.json
{
  "openapi": "3.0.2",
  "paths": {
...
    "/api/v1/posts": {
      "post": {
        "summary": "Create a post or comment",
        "tags": [
          "Post"
        ],
        "parameters": [],
        "operationId": "createPost",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "communityId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "parentPostId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "postType": {
                    "type": "string",
                    "enum": [
                      "comment",
                      "post"
                    ]
                  },
                  "title": {
                    "type": "string",
                    "nullable": true
                  },
                  "content": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": [
                  "postType"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "authorId": {
                      "type": "string"
                    },
                    "communityId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "parentPostId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "postType": {
                      "type": "string",
                      "enum": [
                        "comment",
                        "post"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "nullable": true
                    },
                    "content": {
                      "type": "string",
                      "nullable": true
                    },
                    "upvotes": {
                      "type": "number"
                    },
                    "downvotes": {
                      "type": "number"
                    },
                    "commentCount": {
                      "type": "number"
                    },
                    "engagement": {
                      "type": "number"
                    },
                    "isPopularityUpdated": {
                      "type": "boolean"
                    },
                    "deletedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "createdAt",
                    "updatedAt",
                    "authorId",
                    "communityId",
                    "parentPostId",
                    "postType",
                    "title",
                    "content",
                    "upvotes",
                    "downvotes",
                    "commentCount",
                    "engagement",
                    "isPopularityUpdated",
                    "deletedAt"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/posts/{id}": {
      "patch": {
        "summary": "Update a post or comment",
        "tags": [
          "Post"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "operationId": "updatePost",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "nullable": true
                  },
                  "content": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "authorId": {
                      "type": "string"
                    },
                    "communityId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "parentPostId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "postType": {
                      "type": "string",
                      "enum": [
                        "comment",
                        "post"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "nullable": true
                    },
                    "content": {
                      "type": "string",
                      "nullable": true
                    },
                    "upvotes": {
                      "type": "number"
                    },
                    "downvotes": {
                      "type": "number"
                    },
                    "commentCount": {
                      "type": "number"
                    },
                    "engagement": {
                      "type": "number"
                    },
                    "isPopularityUpdated": {
                      "type": "boolean"
                    },
                    "deletedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "createdAt",
                    "updatedAt",
                    "authorId",
                    "communityId",
                    "parentPostId",
                    "postType",
                    "title",
                    "content",
                    "upvotes",
                    "downvotes",
                    "commentCount",
                    "engagement",
                    "isPopularityUpdated",
                    "deletedAt"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete a post or comment",
        "tags": [
          "Post"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "operationId": "deletePost",
        "responses": {
          "204": {
            "description": "204"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Get a post or comment",
        "tags": [
          "Post"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "operationId": "getPostById",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "authorId": {
                      "type": "string"
                    },
                    "communityId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "parentPostId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "postType": {
                      "type": "string",
                      "enum": [
                        "comment",
                        "post"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "nullable": true
                    },
                    "content": {
                      "type": "string",
                      "nullable": true
                    },
                    "upvotes": {
                      "type": "number"
                    },
                    "downvotes": {
                      "type": "number"
                    },
                    "commentCount": {
                      "type": "number"
                    },
                    "engagement": {
                      "type": "number"
                    },
                    "isPopularityUpdated": {
                      "type": "boolean"
                    },
                    "deletedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "postMedia": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "userId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "postId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "downloadKey": {
                            "type": "string"
                          },
                          "fileSize": {
                            "type": "number"
                          },
                          "mimeType": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "userId",
                          "postId",
                          "downloadKey",
                          "fileSize",
                          "mimeType",
                          "createdAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "id",
                    "createdAt",
                    "updatedAt",
                    "authorId",
                    "communityId",
                    "parentPostId",
                    "postType",
                    "title",
                    "content",
                    "upvotes",
                    "downvotes",
                    "commentCount",
                    "engagement",
                    "isPopularityUpdated",
                    "deletedAt",
                    "postMedia"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "version": "1.0.0",
    "title": "Nexus REST API - Documentation",
    "description": "Nexus REST API - Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "x-service-token": {
        "type": "apiKey",
        "name": "x-service-token",
        "in": "header"
      }
    }
  }
}